+++ src/routes/components/accordions/+page.svelte
