+++ src/routes/components/avatars/+page.svelte
