+++ src/routes/components/badges/+page.svelte
