+++ src/routes/components/buttons/+page.svelte
