+++ src/routes/components/paginators/+page.svelte
