+++ src/routes/components/range-sliders/+page.svelte
