+++ src/routes/components/slide-toggles/+page.svelte
