+++ src/routes/components/steppers/+page.svelte
