+++ src/routes/stores/+server.ts
